Update patch set 1
Patch Set 1:

yeah, i don't know whether to make these changes or not. on the one hand, the TEMP_FAILURE_RETRYs are useless at best and harmful at worst. on the other, people are so used to seeing TEMP_FAILURE_RETRY around system calls that it looks wrong to see ones that aren't wrapped.

i'm pretty sure this isn't the first time we've learned this, which is why i think we have to just fix the C library. but i'm conflicted about whether to touch any of the callers or just leave people with the code they personally think is the more obvious code. (now that all options are correct.)

Patch-set: 1
Label: Code-Review=0