Fix potential crash during system properties tests/benchmarks
prop_info structures returned from the system properties APIs are
meant to be valid for the entire lifetime of a program, however the
system properties tests and benchmarks were invalidating these when
they created a temporary set of system property areas for testing.

This change uses two new internal APIs for storing and restoring the
set of system property areas to maintain the validity of the prop_info
structures.

Bug: 62197783
Test: JniInvocation_test

Change-Id: I20f18cb47dd7e3f358a4e9ae389baf25d021bfd2