Update patch set 1
Patch Set 1:

seems like it was always a hack that <jni.h> was in libnativehelpers. (and on the global include path.) i think exposing this is fine, but i don't think libnativehelper_headers is the right name for this.

i'm not sure what the plan is for cleaning this up in the new build system, or even if they've thought about it yet.

Patch-set: 1
CC: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>