Update patch set 1
Patch Set 1: Code-Review+1

> Patch Set 1: -Code-Review
> 
> I am very not happy with this. To me, this is a bug in libc++. Simple as that.

If anything, it's a bug in the C++ standard, but not something that will get fixed (as noexcept is an ABI-affecting part of the function). Chih-hung already pointed out the comment about why this is still necessary in the bug.

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1