EXPERIMENTAL: C ABI for libnativehelper
Maintain existing C++ API, but export a C ABI.

libnativehelper might live on device or inside the ART APEX. Either
way, if we libc++ is dynamically linked for on device and inside the
ART APEX then we'd need a C ABI between the two layers in case of
libc++ version skew.

This CL attempts to do this, by introducing a curated list of exported
C functions.

Bug: b/110870380
Test: Builds

Change-Id: Idaa4ddd3118cbb4f3997abdcf355220ef80eb775