Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > aren't the ART folks trying to remove this?
> > 
> > you can always do this in your own code, you know...
> 
> We could, but I would like to avoid having many copies of this in case more teams need to use this in the future.

In conflicting work we'd like to reduce libnativehelper's API and remove all of the JNI constants there. libnativehelper is going to be updatable independently of the rest of the system.

The attached bug is a touchscreen problem and doesn't describe why this change is needed.

Patch-set: 1
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>