Update patch set 1
Patch Set 1: -Autosubmit

> Patch Set 1: Code-Review-1
> 
> A single user does not warrant global caching. The ART team (as owner of this) wants to get rid of this global cache (because it's terrible, being shared between instances).

Thanks all for the feedback. This isn't relevant to the bug that I linked, but I was hoping to use this in ag/5195775.

If it makes any difference, there's one more user that I found of this:
https://android.googlesource.com/platform/frameworks/base/+/master/core/jni/android/graphics/Graphics.cpp#765

Is there a replacement for JniConstants available? I'm a bit concerned that this another one of those "deprecated, but the new thing isn't ready yet". Once the replacement is available, wouldn't it be easier to refactor everything at once, in which case, it should help if everyone actually uses JniConstants instead of just copying code for the specific bits they need?

Since this should not be used, can there be a build warning about including the JniConstants header, similar to what is being done for cutils/log currently?

Patch-set: 1
Reviewer: Gerrit User 1149384 <1149384@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=0
Label: Presubmit-Ready=0