Update patch set 3
Patch Set 3: Code-Review+2

> Patch Set 1:
> 
> (I always forget whether the plan still stands to add libnativehelper to the NDK or not, and whether C++11 would be a blocker, so adding folks)

We'd like to, but don't have a good mechanism for distributing it yet (we'd rather ship it as an optional library rather than bundling yet more into the core NDK).

My first thought is that it would be best to continue supporting C++98 for the NDK, but given that it would be an optional library and C++14 is actually the Clang default these days, I think this is probably fine.

Patch-set: 3
Reviewer: Gerrit User 1043845 <1043845@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2