Update patch set 1
Patch Set 1:

This doesn't seem to preserve expected semantics around signalling file/socket closed exceptions. If semantics are preserved by a different means, or it's okay to be transiently unaligned, then it'd be good to have a comment explaining this.

It might be cleaner to just have a single #ifdef block that contains dummy functions for pthread_kill, sigaction, and the signal constant.

Patch-set: 1
CC: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>