Add backword compatibility methods to AsynchronousCloseMonitor.
Step 1 of several. Fixes CTS and verifies backword compatibility.

Older, unbundled versions of Conscrypt assume these C++ symbols exist
and work in this way.  Currently the bundled Conscrypt is using the same
symbols and so CTS is a proxy for the old behaviour.

Next steps: Move this implementation to libjavacore, update Conscrypt
to use the correct C APIs, and add a test to ensure we don't break
compatibility until we're ready.

Bug: 122881655
Bug: 123121395
Test: atest CtsLibcoreOkHttpTestCases:com.squareup.okhttp.CallTest
Change-Id: Ia7206aa5ad5d5d8ae048026a295ac8808ba89296