Add com.android.art.{host|testing} to apex_available
Soong previously created an APEX variant of a module even if the
module's apex_available property doens't contain the apex name. The
check was done only when the module is being included in the APEX.
APEXes for the host and test APEXes were automatically exempted.

However, Soong now doesn't create the apex variant if the module is not
available for the apex. This is to fix a bug that statically linked
modules are not check against apex_available; because they are not seen
as a file in the APEX.

As a result, we need to explicitly add the host and test APEX names in
the apex_available property.

Bug: 147671264
Test: m
Change-Id: Ifaa9e5e48b5b08fccb517233b411eaa009521cf8