apex_available becomes more correct
Previously, Soong didn't follow the static dependencies when checking
apex_available. As a result, a module that is not available to an APEX
(or to the platform) was able to be statically linked from the APEX (or
the platform).

As the Soong bug is fixed, unchecked dependencies have revealed and this
change reflects the finding.

Bug: 147671264
Test: m
Change-Id: Ic0a92cb8db3d784cff7983c53bc26f5bc27b8209